Save draft, Cancel post, and Clear draft
Implemented the Save draft for post a question which will store the question for the user until the user is ready to post it. Cancel button is go back to the main menu. The clear draft is feature I added so if the user wants to ask some other question than the previous one.